Refactored Controls component
