+++ src/components/CreateGroupForm/CreateGroupForm.vue
@@ -1,0 +1,3 @@
+<script src="./CreateGroupForm.js"></script>
+
+<style src="./CreateGroupForm.css"></style>
