+++ src/components/CreateItemForm/CreateItemForm.vue
@@ -1,0 +1,3 @@
+<script src="./CreateItemForm.js"></script>
+
+<style src="./CreateItemForm.css"></style>
