+++ src/components/ManageStateForm/ManageStateForm.vue
@@ -1,0 +1,3 @@
+<script src="./ManageStateForm.js"></script>
+
+<style src="./ManageStateForm.css"></style>
